refactor: TobbleBox -> Box
